+++ tests/unit/common.py
@@ -1,7 +1,7 @@
 import os
 import time
 import inspect
-from abc import ABC
+from abc import ABC, abstractmethod
 from pathlib import Path
 
 import torch
@@ -12,7 +12,7 @@ from torch.multiprocessing import Process
 
 import pytest
 from _pytest.outcomes import Skipped
-from _pytest.fixtures import FixtureLookupError
+from _pytest.fixtures import FixtureLookupError, FixtureFunctionMarker
 
 # Worker timeout *after* the first worker has completed.
 DEEPSPEED_UNIT_WORKER_TIMEOUT = 120
@@ -63,53 +63,42 @@ def set_cuda_visibile():
     os.environ["CUDA_VISIBLE_DEVICES"] = ",".join(dev_id_list)
 
 
-class DistributedTest(ABC):
-    is_dist_test = True
+class DistributedExec(ABC):
+    """
+    Base class for distributed execution of functions/methods. Contains common
+    methods needed for DistributedTest and DistributedFixture.
+    """
     world_size = 2
     backend = "nccl"
     init_distributed = True
     set_dist_env = True
 
-    # Temporary directory that is shared among test methods in a class
-    @pytest.fixture(autouse=True, scope="class")
-    def class_tmpdir(self, tmpdir_factory):
-        fn = tmpdir_factory.mktemp(self.__class__.__name__)
-        return fn
+    @abstractmethod
+    def run(self):
+        ...
 
-    def _run_test(self, request):
-        self.current_test = self._get_current_test_func(request)
-        self.test_kwargs = self._get_test_kwargs(request)
-
-        # Catch world_size override pytest mark
-        for mark in getattr(request.function, "pytestmark", []):
-            if mark.name == "world_size":
-                world_size = mark.args[0]
-                break
-        else:
-            world_size = self.world_size
-
+    def __call__(self, request=None):
+        self._fixture_kwargs = self._get_fixture_kwargs(request, self.run)
+        world_size = self.world_size
         if isinstance(world_size, int):
             world_size = [world_size]
         for procs in world_size:
             self._launch_procs(procs)
             time.sleep(0.5)
 
-    def _get_current_test_func(self, request):
-        # DistributedTest subclasses may have multiple test methods
-        func_name = request.function.__name__
-        return getattr(self, func_name)
-
-    def _get_test_kwargs(self, request):
+    def _get_fixture_kwargs(self, request, func):
+        if not request:
+            return {}
         # Grab fixture / parametrize kwargs from pytest request object
-        test_kwargs = {}
-        params = inspect.getfullargspec(self.current_test).args
+        fixture_kwargs = {}
+        params = inspect.getfullargspec(func).args
         params.remove("self")
         for p in params:
             try:
-                test_kwargs[p] = request.getfixturevalue(p)
+                fixture_kwargs[p] = request.getfixturevalue(p)
             except FixtureLookupError:
                 pass  # test methods can have kwargs that are not fixtures
-        return test_kwargs
+        return fixture_kwargs
 
     def _launch_procs(self, num_procs):
         mp.set_start_method('forkserver', force=True)
@@ -174,7 +163,7 @@ class DistributedTest(ABC):
             torch.cuda.set_device(local_rank)
 
         try:
-            self.current_test(**self.test_kwargs)
+            self.run(**self._fixture_kwargs)
         except BaseException as e:
             if isinstance(e, Skipped):
                 skip_msg.put(e.msg)
@@ -188,107 +177,148 @@ class DistributedTest(ABC):
             dist.destroy_process_group()
 
 
-def distributed_test(world_size=2, backend='nccl'):
-    """A decorator for executing a function (e.g., a unit test) in a distributed manner.
-    This decorator manages the spawning and joining of processes, initialization of
-    deepspeed.comm, and catching of errors.
-
-    Usage example:
-        @distributed_test(worker_size=[2,3])
-        def my_test():
-            rank = dist.get_rank()
-            world_size = dist.get_world_size()
-            assert(rank < world_size)
-
-    Arguments:
-        world_size (int or list): number of ranks to spawn. Can be a list to spawn
-        multiple tests.
+class DistributedFixture(DistributedExec):
     """
-    def dist_wrap(run_func):
-        """Second-level decorator for dist_test. This actually wraps the function. """
-        def dist_init(local_rank, num_procs, *func_args, **func_kwargs):
-            """Initialize deepspeed.comm and execute the user function. """
-            os.environ['MASTER_ADDR'] = '127.0.0.1'
-            os.environ['MASTER_PORT'] = get_master_port()
-            os.environ['LOCAL_RANK'] = str(local_rank)
-            # NOTE: unit tests don't support multi-node so local_rank == global rank
-            os.environ['RANK'] = str(local_rank)
-            os.environ['WORLD_SIZE'] = str(num_procs)
+    Implementation that extends @pytest.fixture to allow for distributed execution.
+    This is primarily meant to be used when a test requires executing two pieces of
+    code with different world sizes.
 
-            # turn off NCCL logging if set
-            os.environ.pop('NCCL_DEBUG', None)
+    There are 2 parameters that can be modified:
+        - world_size: int = 2 -- the number of processes to launch
+        - backend: Literal['nccl','mpi','gloo'] = 'nccl' -- which backend to use
 
-            set_cuda_visibile()
+    Features:
+        - able to call pytest.skip() inside fixture
+        - can be reused by multiple tests
+        - can accept other fixtures as input
 
-            deepspeed.init_distributed(dist_backend=backend)
-            #dist.init_process_group(backend=backend)
-            dist.barrier()
+    Limitations:
+        - cannot use @pytest.mark.parametrize
+        - world_size cannot be modified after definition and only one world_size value is accepted
+        - any fixtures used must also be used in the test that uses this fixture (see example below)
+        - return values cannot be returned. Passing values to a DistributedTest
+          object can be achieved using class_tmpdir and writing to file (see example below)
 
-            if torch.cuda.is_available():
-                torch.cuda.set_device(local_rank)
+    Usage:
+        - must implement a run(self, ...) method
+        - fixture can be used by making the class name input to a test function
 
-            run_func(*func_args, **func_kwargs)
+    Example:
+        @pytest.fixture(params=[10,20])
+        def regular_pytest_fixture(request):
+            return request.param
 
-            # make sure all ranks finish at the same time
-            dist.barrier()
-            # tear down after test completes
-            dist.destroy_process_group()
+        class distributed_fixture_example(DistributedFixture):
+            world_size = 4
 
-        def dist_launcher(num_procs, *func_args, **func_kwargs):
-            """Launch processes and gracefully handle failures. """
+            def run(self, regular_pytest_fixture, class_tmpdir):
+                assert int(os.environ["WORLD_SIZE"]) == self.world_size
+                local_rank = os.environ["LOCAL_RANK"]
+                print(f"Rank {local_rank} with value {regular_pytest_fixture}")
+                with open(os.path.join(class_tmpdir, f"{local_rank}.txt"), "w") as f:
+                    f.write(f"{local_rank},{regular_pytest_fixture}")
 
-            # Spawn all workers on subprocesses.
-            processes = []
-            for local_rank in range(num_procs):
-                p = Process(target=dist_init,
-                            args=(local_rank,
-                                  num_procs,
-                                  *func_args),
-                            kwargs=func_kwargs)
-                p.start()
-                processes.append(p)
+        class TestExample(DistributedTest):
+            world_size = 1
 
-            # Now loop and wait for a test to complete. The spin-wait here isn't a big
-            # deal because the number of processes will be O(#GPUs) << O(#CPUs).
-            any_done = False
-            while not any_done:
-                for p in processes:
-                    if not p.is_alive():
-                        any_done = True
-                        break
+            def test(self, distributed_fixture_example, regular_pytest_fixture, class_tmpdir):
+                assert int(os.environ["WORLD_SIZE"]) == self.world_size
+                for rank in range(4):
+                    with open(os.path.join(class_tmpdir, f"{rank}.txt"), "r") as f:
+                        assert f.read() == f"{rank},{regular_pytest_fixture}"
+    """
+    is_dist_fixture = True
 
-            # Wait for all other processes to complete
-            for p in processes:
-                p.join(DEEPSPEED_UNIT_WORKER_TIMEOUT)
+    # These values are just placeholders so that pytest recognizes this as a fixture
+    _pytestfixturefunction = FixtureFunctionMarker(scope="function", params=None)
+    __name__ = ""
 
-            failed = [(rank, p) for rank, p in enumerate(processes) if p.exitcode != 0]
-            for rank, p in failed:
-                # If it still hasn't terminated, kill it because it hung.
-                if p.exitcode is None:
-                    p.terminate()
-                    pytest.fail(f'Worker {rank} hung.', pytrace=False)
-                if p.exitcode < 0:
-                    pytest.fail(f'Worker {rank} killed by signal {-p.exitcode}',
-                                pytrace=False)
-                if p.exitcode > 0:
-                    pytest.fail(f'Worker {rank} exited with code {p.exitcode}',
-                                pytrace=False)
+    def __init__(self):
+        assert isinstance(self.world_size, int), "Only one world size is allowed for distributed fixtures"
+        self.__name__ = type(self).__name__
+        _pytestfixturefunction = FixtureFunctionMarker(scope="function",
+                                                       params=None,
+                                                       name=self.__name__)
 
-        def run_func_decorator(*func_args, **func_kwargs):
-            """Entry point for @distributed_test(). """
 
-            if isinstance(world_size, int):
-                dist_launcher(world_size, *func_args, **func_kwargs)
-            elif isinstance(world_size, list):
-                for procs in world_size:
-                    dist_launcher(procs, *func_args, **func_kwargs)
-                    time.sleep(0.5)
-            else:
-                raise TypeError(f'world_size must be an integer or a list of integers.')
+class DistributedTest(DistributedExec):
+    """
+    Implementation for running pytest with distributed execution.
 
-        return run_func_decorator
+    There are 2 parameters that can be modified:
+        - world_size: Union[int,List[int]] = 2 -- the number of processes to launch
+        - backend: Literal['nccl','mpi','gloo'] = 'nccl' -- which backend to use
 
-    return dist_wrap
+    Features:
+        - able to call pytest.skip() inside tests
+        - works with pytest fixtures, parametrize, mark, etc.
+        - can contain multiple tests (each of which can be parametrized separately)
+        - class methods can be fixtures (usable by tests in this class only)
+        - world_size can be changed for individual tests using @pytest.mark.world_size(world_size)
+        - class_tmpdir is a fixture that can be used to get a tmpdir shared among
+          all tests (including DistributedFixture)
+
+    Usage:
+        - class name must start with "Test"
+        - must implement one or more test*(self, ...) methods
+
+    Example:
+        @pytest.fixture(params=[10,20])
+        def val1(request):
+            return request.param
+
+        @pytest.mark.fast
+        @pytest.mark.parametrize("val2", [30,40])
+        class TestExample(DistributedTest):
+            world_size = 2
+
+            @pytest.fixture(params=[50,60])
+            def val3(self, request):
+                return request.param
+
+            def test_1(self, val1, val2, str1="hello world"):
+                assert int(os.environ["WORLD_SIZE"]) == self.world_size
+                assert all(val1, val2, str1)
+
+            @pytest.mark.world_size(1)
+            @pytest.mark.parametrize("val4", [70,80])
+            def test_2(self, val1, val2, val3, val4):
+                assert int(os.environ["WORLD_SIZE"]) == 1
+                assert all(val1, val2, val3, val4)
+    """
+    is_dist_test = True
+
+    # Temporary directory that is shared among test methods in a class
+    @pytest.fixture(autouse=True, scope="class")
+    def class_tmpdir(self, tmpdir_factory):
+        fn = tmpdir_factory.mktemp(self.__class__.__name__)
+        return fn
+
+    def run(self, **fixture_kwargs):
+        self._current_test(**fixture_kwargs)
+
+    def __call__(self, request):
+        self._current_test = self._get_current_test_func(request)
+        self._fixture_kwargs = self._get_fixture_kwargs(request, self._current_test)
+
+        # Catch world_size override pytest mark
+        for mark in getattr(request.function, "pytestmark", []):
+            if mark.name == "world_size":
+                world_size = mark.args[0]
+                break
+        else:
+            world_size = self.world_size
+
+        if isinstance(world_size, int):
+            world_size = [world_size]
+        for procs in world_size:
+            self._launch_procs(procs)
+            time.sleep(0.5)
+
+    def _get_current_test_func(self, request):
+        # DistributedTest subclasses may have multiple test methods
+        func_name = request.function.__name__
+        return getattr(self, func_name)
 
 
 def get_test_path(filename):
