+++ tests/unit/runtime/test_multi_output_model.py
@@ -1,46 +1,38 @@
 import torch
 import deepspeed
 from pytest import approx
-from .common import distributed_test
-from .simple_model import args_from_dict
-from .multi_output_model import MultiOutputModel, multi_output_dataloader
+from unit.common import DistributedTest
+from unit.multi_output_model import MultiOutputModel, multi_output_dataloader
 
 
-def create_config_dict(micro_batch_size, grad_accumulation_steps, world_size):
-    return {
-        "train_micro_batch_size_per_gpu": micro_batch_size,
-        "gradient_accumulation_steps": grad_accumulation_steps,
-        "train_batch_size": micro_batch_size * grad_accumulation_steps * world_size,
-        "steps_per_print": 1,
-        "optimizer": {
-            "type": "Adam",
-            "params": {
-                "lr": 0.00015
-            }
-        },
-        "fp16": {
-            "enabled": True
-        }
-    }
-
-
-def test_two_output_model(tmpdir):
-    gradient_accumulation_steps = 2
-    micro_batch_size = 1
+class TestTwoOutputModel(DistributedTest):
     world_size = 1
-    config_dict = create_config_dict(micro_batch_size,
-                                     gradient_accumulation_steps,
-                                     world_size)
 
-    hidden_dim = 10
-    weight_value = 0.1
-    args = args_from_dict(tmpdir, config_dict)
+    def test(self, tmpdir):
+        grad_accumulation_steps = 2
+        micro_batch_size = 1
+        world_size = self.world_size
+        config_dict = {
+            "train_micro_batch_size_per_gpu": micro_batch_size,
+            "gradient_accumulation_steps": grad_accumulation_steps,
+            "train_batch_size": micro_batch_size * grad_accumulation_steps * world_size,
+            "steps_per_print": 1,
+            "optimizer": {
+                "type": "Adam",
+                "params": {
+                    "lr": 0.00015
+                }
+            },
+            "fp16": {
+                "enabled": True
+            }
+        }
 
-    model = MultiOutputModel(hidden_dim, weight_value)
+        hidden_dim = 10
+        weight_value = 0.1
 
-    @distributed_test(world_size=[1])
-    def _test_two_output_model(args, model, hidden_dim):
-        model, _, _, _ = deepspeed.initialize(args=args,
+        model = MultiOutputModel(hidden_dim, weight_value)
+        model, _, _, _ = deepspeed.initialize(config=config_dict,
                                               model=model,
                                               model_parameters=model.parameters())
         total_samples = 4
@@ -69,35 +61,44 @@ def test_two_output_model(tmpdir):
 
             summed_loss = sum(loss_tuple)
             scaled_loss = model.backward(summed_loss)
-            expected_scaled_loss = summed_loss.float() / gradient_accumulation_steps
+            expected_scaled_loss = summed_loss.float() / grad_accumulation_steps
             assert scaled_loss.item() == approx(expected_scaled_loss.item())
 
             model.step()
 
-    _test_two_output_model(args=args, model=model, hidden_dim=hidden_dim)
 
-
-def test_three_output_model(tmpdir):
-    gradient_accumulation_steps = 3
-    micro_batch_size = 1
+class TestThreeOutputModel(DistributedTest):
     world_size = 1
-    config_dict = create_config_dict(micro_batch_size,
-                                     gradient_accumulation_steps,
-                                     world_size)
 
-    hidden_dim = 10
-    weight_value = 0.1
-    args = args_from_dict(tmpdir, config_dict)
+    def test(self, tmpdir):
+        grad_accumulation_steps = 3
+        micro_batch_size = 1
+        world_size = 1
+        config_dict = {
+            "train_micro_batch_size_per_gpu": micro_batch_size,
+            "gradient_accumulation_steps": grad_accumulation_steps,
+            "train_batch_size": micro_batch_size * grad_accumulation_steps * world_size,
+            "steps_per_print": 1,
+            "optimizer": {
+                "type": "Adam",
+                "params": {
+                    "lr": 0.00015
+                }
+            },
+            "fp16": {
+                "enabled": True
+            }
+        }
 
-    model = MultiOutputModel(hidden_dim, weight_value)
+        hidden_dim = 10
+        weight_value = 0.1
 
-    @distributed_test(world_size=[1])
-    def _test_three_output_model(args, model, hidden_dim):
-        model, _, _, _ = deepspeed.initialize(args=args,
+        model = MultiOutputModel(hidden_dim, weight_value)
+        model, _, _, _ = deepspeed.initialize(config=config_dict,
                                               model=model,
                                               model_parameters=model.parameters())
 
-        total_samples = gradient_accumulation_steps * micro_batch_size * 2
+        total_samples = grad_accumulation_steps * micro_batch_size * 2
         data_loader = multi_output_dataloader(model=model,
                                               total_samples=total_samples,
                                               hidden_dim=hidden_dim,
@@ -127,9 +128,7 @@ def test_three_output_model(tmpdir):
 
             summed_loss = sum(loss_tuple)
             scaled_loss = model.backward(summed_loss)
-            expected_scaled_loss = summed_loss.float() / gradient_accumulation_steps
+            expected_scaled_loss = summed_loss.float() / grad_accumulation_steps
             assert scaled_loss.item() == approx(expected_scaled_loss.item())
 
             model.step()
-
-    _test_three_output_model(args=args, model=model, hidden_dim=hidden_dim)
